+++ backend/src/Migrations/Version20220527135157.php
@@ -1,0 +1,22 @@
+<?php
+/** @noinspection PhpUnused */
+
+declare(strict_types=1);
+
+namespace Neucore\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+final class Version20220527135157 extends AbstractMigration
+{
+    public function up(Schema $schema): void
+    {
+        $this->addSql('ALTER TABLE esi_tokens ADD last_checked DATETIME DEFAULT NULL');
+    }
+
+    public function down(Schema $schema): void
+    {
+        $this->addSql('ALTER TABLE esi_tokens DROP last_checked');
+    }
+}